+++ Source/dsp/DiodeNLEQ.h
@@ -16,11 +16,11 @@ class DiodeNLEQ : public NonLinearEquationBase
 {
 public:
     DiodeNLEQ();
-    float calculateCurrents(Eigen::MatrixXd voltages, int vn_index, int function_index) override;
+    double calculateCurrents(Eigen::MatrixXd voltages, int vn_index, int function_index) override;
     void calculateJacobian(Eigen::MatrixXd* jacobian, Eigen::MatrixXd voltages, int vn_index) override;
 
 private:
-    float Is = 1e-15;
-    float Vt = 25.85e-3;
-    float eta = 1;
+    double Is = 1e-15;
+    double Vt = 25.85e-3;
+    double eta = 1;
 };
