+++ Source/dsp/NonLinearEquationBase.h
@@ -17,7 +17,7 @@ class NonLinearEquationBase
 public:
 
     virtual ~NonLinearEquationBase() = default;
-    virtual float calculateCurrents(Eigen::MatrixXd voltages, int vn_index, int function_index) { return -666.0f; }
+    virtual double calculateCurrents(Eigen::MatrixXd voltages, int vn_index, int function_index) { return -666.0; }
     virtual void calculateJacobian(Eigen::MatrixXd* jacobian, Eigen::MatrixXd voltages, int vn_index) {}
     int getNumberOfFunctions() { return numberOfFunctions; };
 protected:
